+++ __init__.py
@@ -20,7 +20,7 @@ bl_info = {
 	'name': 'BsMax',
 	'description': 'Package of many tools + other CG apps UI simulator',
 	'author': 'Naser Merati (Nevil)',
-	'version': (0, 1, 0, 20220606),
+	'version': (0, 1, 0, 20220621),
 	'blender': (2, 80, 0),# 2.80 ~ 3.2
 	'location': 'Almost Everywhere in Blender',
 	'wiki_url': 'https://github.com/NevilArt/BsMax/wiki',